# Use cython to speed-up wrap angle