--- a/docs/changes/coordinates/16326.feature.rst
+++ b/docs/changes/coordinates/16326.feature.rst
@@ -1,0 +1,1 @@
+Add new cython module to optimize angle wrapping.