@maxnoe 2024-04-23
Here is a plot of the improvements, generated using the code below
![wrap_at_performance](https://github.com/astropy/astropy/assets/5488440/69242e35-0613-428f-b863-86ff0f374749)



```
from pathlib import Path
import astropy.units as u
from astropy.units import Quantity
import numpy as np
from astropy.coordinates.angles import Longitude, Latitude, Angle
import json

rng = np.random.default_rng(0)

shapes = (1, 100, 10000, (500, 500))

p_nan = (0, 1e-2, 0.1, 0.5)
wrap = Angle(180, u.deg)

results = []

for shape in shapes:
    for p in p_nan:
        values = rng.uniform(-720, 720, np.prod(shape))
        if p > 0:
            values[rng.choice(values.size, int(np.ceil(p * values.size)), replace=False)] = np.nan

        values = values.reshape(shape)
        angle = Angle(values, u.deg)

        print('total', shape, 'nans', np.count_nonzero(np.isnan(values)))
        result = get_ipython().run_line_magic("timeit", "-o angle.wrap_at(wrap)")
        results.append([{"shape": shape, "p_nan": p, "mean": result.average, "std": result.stdev}])

Path("performance.json").write_text(json.dumps(results))
```