@maxnoe 2024-04-23
The test failure *looks* unrelated at first glance... any idea?

https://github.com/astropy/astropy/actions/runs/8798691509/job/24146244116?pr=16326