@neutrinoceros 2024-04-23
Looks like macOS runners were just migrated to AMR64 and we hit #16319 and #16320, so yeah, definitely unrelated to this PR