@pllim 2024-04-23
> I'd also suggest to keep open https://github.com/astropy/astropy/pull/13497 for the time being

That is a merged PR, there is nothing to keep open. Was there a typo?