@pllim 2024-04-23
Needs:

* change log
* fix pre-commit check
* identify a coordinates maintainer that is well-versed in Cython to review