@pllim 2024-04-23
> Looks like macOS runners were just migrated to ARM64

We can skip those tests in OSX or downgrade that job to older mac x86 runners. But we can discuss in a different issue.