@pllim 2024-04-23
Re: `benchmark` label -- I added it. When it completes, you will see a summary in the log and also option to download artifacts.