@pllim 2024-04-23
Hmm is this related or just VM being flaky?

```
| Change   | Before [e9ba0dba]    | After [9eb6abd7]    |   Ratio | Benchmark (Parameter)          |
|----------|----------------------|---------------------|---------|--------------------------------|
| +        | 44.1±0.1μs           | 84.6±40μs           |    1.92 | units.time_quantity_times_unit |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
Error: PERFORMANCE DECREASED.
```