@maxnoe 2024-04-23
> 84.6±40μs   

The uncertainty is 50 % of the value...