@maxnoe 2024-04-23
Now I am very confused... I checked also creating an `AltAz` instance, which was our original bottleneck, and that is slower here than on main... how can that be if the function itself is faster ...