@maxnoe 2024-04-23
Scratch that, I must have fooled myself somehow with switching branches and editable installs...