@astrofrog 2024-04-24
I wonder if the cython wrapping function could be more generally useful in other parts of astropy - for example for folding time series. Perhaps we should think of whether it can be written with more general variable names and placed in e.g. utils or similar, for use in different sub-packages? (only if that generalization does not impact performance).