@saimn 2024-04-24
Please add a `setup_package.py` file to define NPY_NO_DEPRECATED_API and avoid NPY_1_7_API_VERSION warnings (see other subpackages for examples).