@saimn 2024-04-24
I'm not expert with `Angle` but while doing quick tests to play with cython I'm just wondering if we need ufunc/fused types: 
- from what I can see `Angle` always converts values to float64 (which makes sense for physical units such as angles), so why supporting int variants ?
- `wrap_angle` and `period` are scalar so no need for array/broadcast etc. 
- seems in that case that a simple loop on a memoryview would be faster than the ufunc without needing the extra features (and smaller .so file I guess)