@astrofrog 2024-04-24
For fun: if we don't think we need the fused dtypes, we could also just use a plain C extension which compiles and runs faster: https://gist.github.com/astrofrog/136d15ebd650b4465060e314fccd7988 - runs about twice as fast (as Cython) for large arrays and about 10x faster for small arrays. In practice actually need to do the whole thing with the outer/inner loop to be able to deal with non-contiguous arrays and endian etc but just thought I'd try it out as an experiment (I did not actually write a single line of that module, it was an experiment with AI-generated code!). If this really is a bottleneck in coordinates, it might be worth writing the C ourselves - in a number of projects I've worked on before I've found you can often typically get 2x better performance when dealing with large arrays.