@mhvk 2024-04-30
I have looked at the cython implementations and find them quite simple, which is good, but still they make the code less clear and add some 100kiB of binary library. I wonder whether the speed benefits are actually worth it for the larger question of wanting to speed up coordinate transformations, etc. In particular, from #13479, just initiating an `Angle` is already substantially slower than a `Quantity`, which really seems rather odd (though that is probably best discussed there).