@dhomeier 2024-06-10
@mhvk are your doubts about putting in the extra effort to rewrite this in C, or for the present cython code?
If I interpret the plot at the top right, this optimisation _will_ already outweigh the overhead of initialising an `Angle` for sufficiently large shapes, so could result in real speedups for many cases.